Move config loader functions from test_setup.py to config.py and update references
