+++ ops/ccs-ops-misc/load_test/common/test_setup.py
@@ -6,41 +6,6 @@ from typing import Dict
 from common import config
 from locust.main import main
 
-from common.stats import StatsStorageConfig
-
-
-def get_client_cert() -> str:
-    '''Checks the config file for the client cert value.
-    '''
-
-    config_file = config.load()
-    return config_file["clientCertPath"]
-
-
-def load_server_public_key() -> str:
-    '''Load the public key to verify the BFD Server's responses or else ignore the warnings from
-    the self-signed cert.
-    '''
-
-    try:
-        config_file = config.load()
-        server_public_key = config_file["serverPublicKey"]
-        return server_public_key if server_public_key else False
-    except KeyError:
-        return False
-
-def load_stats_storage_config() -> StatsStorageConfig:
-    """Load the storage configuration for storing aggregated statistics.
-
-    Returns:
-        StatsStorageConfig: A dataclass representing the storage configuration for aggregated statistics
-    """
-
-    # TODO: This may not be the best spot for this method, since it's used during teardown not setup.
-    # Maybe add a new test_teardown.py, or move the loading of values from config to config.py?
-    config_file = config.load()
-    return config_file["storeStats"]
-
 
 def set_locust_env(config_file: Dict[str, str]):
     '''Sets a number of locust variables needed to run the test.
